tracing: mv DipoleFields -> emfields.dipole
